AI Assistant/requirements.txt
huggingface_hub==0.20.3
langchain==0.1.11
langchain_community==0.0.26
langchain_openai==0.0.8
langchain-experimental==0.0.51
pandas==2.2.1
python-dotenv==1.0.1
streamlit==1.31.1
torch==2.2.0
openpyxl==3.1.2
openai==1.12.0
rapidocr-onnxruntime==1.3.11
gpt4all==2.2.1
llama-cpp-python==0.2.55
langchainhub==0.1.15
sentence_transformers==2.5.1
huggingface_hub

# In case the model does not find the GPU of the machine, do the following:
    # pip uninstall llama-cpp-python
    # python -m pip install llama-cpp-python --prefer-binary --no-cache-dir --extra-index-url=https://jllllll.github.io/llama-cpp-python-cuBLAS-wheels/AVX2/cu122